resolved review issues
