+++ package.json
@@ -31,8 +31,8 @@
     "@types/expect": "24.3.0",
     "@types/jsdom": "16.2.14",
     "@types/mocha": "9.1.0",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
     "@types/react-router-dom": "5.3.3",
     "@types/react-test-renderer": "17.0.1",
     "@typescript-eslint/eslint-plugin": "5.11.0",
@@ -43,6 +43,7 @@
     "copy-webpack-plugin": "10.2.4",
     "cross-env": "7.0.3",
     "css-loader": "6.6.0",
+    "css-minimizer-webpack-plugin": "^3.4.1",
     "eslint": "8.8.0",
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-prettier": "8.3.0",
