+++ src/pages/auth/Login/Login.types.ts
@@ -1,0 +1,8 @@
+import {
+  IMessage,
+  StringOnlyValues,
+} from '../../../components/ConsoleStrategy/ConsoleStrategy.types';
+
+export interface LoginMessages {
+  messages: IMessage<StringOnlyValues>[];
+}
