+++ src/styles/colors.scss
@@ -10,6 +10,7 @@ $light-black-color: #131313;
 //COLORFUL COLORS
 $green-color: #5e8f2d;
 $pink-color: #ff6188;
+$error-color: red;
 $light-blue-color: #78dce8;
 $yellow-color: #fce566;
