+++ src/styles/root.scss
@@ -10,5 +10,4 @@ body {
   & > div:first-child {
     height: 100%;
   }
-
 }
