+++ webpack.config.js
@@ -4,6 +4,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserWebpackPlugin = require('terser-webpack-plugin');
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
 
 const isDev = process.env.NODE_ENV === 'development';
 const isProd = !isDev;
@@ -16,7 +17,7 @@ const optimization = () => {
   };
 
   if (isProd) {
-    config.minimizer = [new TerserWebpackPlugin()];
+    config.minimizer = [new TerserWebpackPlugin(), new CssMinimizerPlugin()];
   }
 
   return config;
@@ -25,9 +26,10 @@ const optimization = () => {
 const filename = (ext) => (isDev ? `[name].${ext}` : `[name].[hash].${ext}`);
 
 module.exports = {
+  context: path.join(__dirname, 'src'),
   mode: 'development',
   entry: {
-    main: ['@babel/polyfill', './src/index.tsx'],
+    main: ['@babel/polyfill', './index.tsx'],
   },
   output: {
     path: path.join(__dirname, '/dist'),
@@ -43,12 +45,11 @@ module.exports = {
   devServer: {
     port: 3000,
     hot: isDev,
-    historyApiFallback: true,
   },
   devtool: isDev ? 'source-map' : false,
   plugins: [
     new HtmlWebpackPlugin({
-      template: './static/index.html',
+      template: '../static/index.html',
       minify: {
         collapseWhitespace: isProd,
       },
@@ -57,7 +58,7 @@ module.exports = {
     new CopyWebpackPlugin({
       patterns: [
         {
-          from: path.join(__dirname, './static/favicon.svg'),
+          from: path.join(__dirname, 'static/favicon.svg'),
           to: path.join(__dirname, 'dist'),
         },
       ],
